WIP: Moving network utilities to a base and implementation classes TODO: balena_network_utility.py may or may not work with wifi-connect hotspot to setup. Investigate further
