+++ device/utilities/network/balena_network_utility.py
@@ -1,18 +1,33 @@
 import subprocess
+from pathlib import Path
 from typing import List, Dict
+from django.conf import settings
+import os
 
-from device.utilities.network.network import NetworkUtility
+from device.utilities.network.base_network_utility import NetworkUtility
+
 
 class BalenaNetworkUtility(NetworkUtility):
     def __init__(self):
         super().__init__()
-        self.wifi_connect_commad = "wifi-connect"
+        self.wifi_connect_commad = "/usr/src/app/wifi-connect"
         self.logger.info("BalenaNetworkUtility: __init__")
+        self.NET_CONFIGURED = settings.DATA_PATH + "/network.configured"
+        self.AP_NAME = "PFC-" + os.getenv("RESIN_DEVICE_NAME_AT_INIT")
+
+    # Override is_connected() to check to see if we need to start up wifi-connect (first boot)
+
+    def is_connected(self) -> bool:
+        base_connected = super().is_connected()
+        # check for network_configured flag.
+        if not base_connected and not Path(self.NET_CONFIGURED).exists():
+            self._start_wifi_connect(True)  # Clear out old config anyway
+            Path(self.NET_CONFIGURED).touch()
+        return base_connected
 
     def get_wifi_ssids(self, exclude_hidden: bool = True, exclude_beaglebones: bool = True) -> List[Dict[str, str]]:
         self.logger.info("BalenaNetworkUtility: get_wifi_ssids: passing")
-        pass
-        #return super().get_wifi_ssids(exclude_hidden, exclude_beaglebones)
+        return []
 
     def join_wifi(self, ssid: str, password: str) -> None:
         self.logger.info("BalenaNetworkUtility: join_wifi: calling _start_wifi_connect()")
@@ -22,7 +37,6 @@ class BalenaNetworkUtility(NetworkUtility):
                           identity: str, phase2: str) -> None:
         self.logger.info("BalenaNetworkUtility: join_wifi_advanced: calling _start_wifi_connect()")
         return self._start_wifi_connect()
-        #super().join_wifi_advance(ssid_name, passphrase, hidden_ssid, security, eap, identity, phase2)
 
     def delete_wifis(self) -> None:
         self.logger.info("BalenaNetworkUtility: delete_wifis: calling _start_wifi_connect(True)")
@@ -30,9 +44,9 @@ class BalenaNetworkUtility(NetworkUtility):
 
     def _start_wifi_connect(self, disconnect: bool = False) -> None:
         self.logger.info("BalenaNetworkUtility in _start_wifi_connect()")
-        clear_flag = "--clear=false"
         if disconnect:
             self.logger.debug("Disconnecting from current wifi")
-            clear_flag = "--clear=true"
-        subprocess.call([self.wifi_connect_commad, clear_flag])
+        # Path(self.NET_CONFIGURED).unlink()
+        subprocess.call(["scripts/platform/reset_balena_app.sh"])
+        # subprocess.call([self.wifi_connect_commad, "-s", self.AP_NAME, "-o", "8765", "-u", "/usr/src/app/ui"])
         return
